requirements.txt
streamlit>=1.28.0
torch>=2.2.0
torchvision>=0.17.0
matplotlib>=3.7.0
numpy>=1.24.0
Pillow>=10.0.0 